pyproject.toml
[project]
name = "ai-agents-pure-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "openai>=1.99.3",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
